+++ Adif Test App/Program.cs
@@ -17,11 +17,19 @@ namespace Adif_Test_App
         static void Main(string[] args)
         {
 
+
             HamRadioDeluxeDatabase.ConnectToDatabase("192.168.64.200", 3307, "PA1REG", "dC7K68zUdGx6FKEu", "PA1REG");
             HamRadioDeluxeDatabase
+
+
+            //HamRadioDeluxeDatabase.ConnectToDatabase("localhost", 3307, "PA1REG", "MyPassword", "PA1REG");
+
             //HamRadioDeluxeDatabase.DatabaseInfo();
             //Console.ReadLine();
             //// CsvUtils.ReadBandList();
+
+          // CsvUtils.ReadBandList();
+
 
             //CsvUtils.CsvReaderQso(@"Z:\Repos\HamRadioLibrary\src\eQsl.ccWithComment.csv",true,true,false);