+++ HamRadioLibrary/PowerShell/HamRadioLibrary.psm1
@@ -154,6 +154,7 @@ function Start-HrdSwlContacts {
 }
 
 
+
 function Get-XlsReportQsoYearMode {
   [CmdletBinding()]
   param
@@ -201,3 +202,7 @@ function Get-XlsReportQsoYearMode {
 Export-ModuleMember -function Start-HrdQsoCallBandMode
 Export-ModuleMember -function Start-HrdSwlContacts
 Export-ModuleMember -function Get-XlsReportQsoYearMode
+
+Export-ModuleMember -function Start-HrdQsoCallBandMode
+Export-ModuleMember -function Start-HrdSwlContacts
+