+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+
 ## [0.0.7] - 2021-02-12
 
 ### Added
@@ -22,6 +23,8 @@ Get-XlsReportQsoYearMode new function.
 ### Changed
 -
 
+
+=======
 
 ## [0.0.5] - 2021-01-25